fix isar, int overflow, stopself when task removed

+++ packages/player/android/src/main/kotlin/br/com/suamusica/player/Media.kt
@@ -3,10 +3,10 @@ package br.com.suamusica.player
 import com.google.gson.annotations.SerializedName
 
 data class Media(
-    @SerializedName("id") val id: Int,
+    @SerializedName("id") val id: Long,
     @SerializedName("name") val name: String,
     @SerializedName("ownerId") val ownerId: Int,
-    @SerializedName("albumId") val albumId: Int,
+    @SerializedName("albumId") val albumId: Long,
     @SerializedName("albumTitle") val albumTitle: String,
     @SerializedName("author") val author: String,
     @SerializedName("url") val url: String,
@@ -15,7 +15,7 @@ data class Media(
     @SerializedName("bigCover") val bigCoverUrl: String,
     @SerializedName("is_verified") val isVerified: Boolean,
     @SerializedName("shared_url") val shareUrl: String,
-    @SerializedName("playlist_id") val playlistId: Int,
+    @SerializedName("playlist_id") val playlistId: Long,
     @SerializedName("is_spot") val isSpot: Boolean,
     @SerializedName("isFavorite") val isFavorite: Boolean?,
     @SerializedName("fallbackUrl") val fallbackUrl: String,
